done with mirage in a more cleaner way
