+++ mirage/helpers/parse-request.js
@@ -1,0 +1,9 @@
+/**
+ * This function returns data that User has posted.
+ * 
+ * @param {Object} request 
+ * @returns Object
+ */
+ export default function getRequestData(request) {
+    return JSON.parse(request.requestBody).data;
+}
