+++ mirage/route-handlers/activity.js
@@ -1,0 +1,5 @@
+export function register(server, ctx) {
+    server.get('/activity', (schema) => {
+        return schema.activities.all();
+    });
+}
