+++ mirage/route-handlers/chatroom.js
@@ -1,0 +1,2 @@
+export function register(server, ctx) {
+}
