+++ mirage/route-handlers/savedsearch.js
@@ -1,0 +1,5 @@
+export function register(server, ctx) {
+    server.get('/savedsearch', (schema) => {
+        return schema.savedsearches.all();
+    });
+}
