+++ mirage/route-handlers/tagged.js
@@ -1,0 +1,2 @@
+export function register(server, ctx) {
+}
