+++ mirage/route-handlers/user.js
@@ -1,0 +1,24 @@
+import pushObjectInModel from '../helpers/push-object-in-model';
+import getValueFromQuery from '../helpers/get-value-from-query';
+
+export function register(server, ctx) {
+    server.get('/user', (schema, request) => {
+        let model = schema.users.all();
+        let userQuery = request.queryParams.query;
+        let id = getValueFromQuery(`User.id`, userQuery);
+        if (id) {
+            pushObjectInModel(model, schema.users.find(id));
+        }
+        return model;
+    });
+
+    server.get('/user/:id', (schema, request) => {
+        let id = request.params.id;
+        let s = server;
+        if (id === "me") {
+            id = ctx.get('currentUser').id
+        }
+        let model = schema.users.find(id);
+        return model;
+    });
+}
