+++ mirage/route-handlers/wiki.js
@@ -1,0 +1,2 @@
+export function register(server, ctx) {
+}
